Fix sizes in IControl_Type

+++ SDLLIB/include/tile.h
@@ -37,6 +37,8 @@
 #ifndef TILE_H
 #define TILE_H
 
+#include <stdint.h>
+
 /*=========================================================================*/
 /* The following prototypes are for the file: ICONSET.CPP						*/
 /*=========================================================================*/
@@ -59,14 +61,14 @@ typedef struct {
 	short	Allocated;		// Was this iconset allocated?
 	short MapWidth;		// Width of map (in icons).
 	short MapHeight;		// Height of map (in icons).
-	long	Size;				// Size of entire iconset memory block.
-	long	Icons;			// Offset from buffer start to icon data.
+	int32_t	Size;				// Size of entire iconset memory block.
+	int32_t	Icons;			// Offset from buffer start to icon data.
 //	unsigned char * Icons;	// Offset from buffer start to icon data.
-	long	Palettes;		// Offset from buffer start to palette data.
-	long	Remaps;			// Offset from buffer start to remap index data.
-	long	TransFlag;		// Offset for transparency flag table.
-	long	ColorMap;		// Offset for color control value table.
-	long	Map;				// Icon map offset (if present).
+	int32_t	Palettes;		// Offset from buffer start to palette data.
+	int32_t	Remaps;			// Offset from buffer start to remap index data.
+	int32_t	TransFlag;		// Offset for transparency flag table.
+	int32_t	ColorMap;		// Offset for color control value table.
+	int32_t	Map;				// Icon map offset (if present).
 //	unsigned char * Map;				// Icon map offset (if present).
 } IControl_Type;
